requirements.txt
streamlit
openai
byaldi
torch
pypandoc
gtts
Pillow
dotenv
deep-translator
sentence-transformers
scikit-learn